requirements.txt
# PQ-V2G Requirements
# Core Python dependencies for the Quantum-Safe V2G system

# Cryptography and Post-Quantum Crypto
cryptography>=41.0.0
pycryptodome>=3.18.0
liboqs-python>=0.8.0  # Post-quantum crypto library
pyoqs>=0.1.0          # Alternative PQ crypto bindings

# Network and Protocol Support
websockets>=11.0.0    # OCPP WebSocket support
aiohttp>=3.8.0        # Async HTTP client/server
requests>=2.31.0      # HTTP requests
socket.io>=5.8.0      # Real-time communication
asyncio>=3.11.0       # Async programming

# TLS and Certificate Handling
pyOpenSSL>=23.0.0     # OpenSSL bindings
certifi>=2023.5.7     # Certificate validation
asn1crypto>=1.5.1     # ASN.1 parsing
pyasn1>=0.5.0         # ASN.1 types

# ISO 15118 and OCPP Protocol Support
xmlschema>=2.3.0      # XML schema validation
lxml>=4.9.0           # XML processing
jsonschema>=4.17.0    # JSON schema validation
pydantic>=2.0.0       # Data validation

# Message Serialization
protobuf>=4.23.0      # Protocol Buffers
msgpack>=1.0.5        # MessagePack serialization
cbor2>=5.4.0          # CBOR serialization

# Database and Storage
SQLAlchemy>=2.0.0     # Database ORM
sqlite3>=3.40.0       # Embedded database
redis>=4.5.0          # In-memory data store

# Logging and Monitoring
structlog>=23.1.0     # Structured logging
prometheus_client>=0.16.0  # Metrics collection
psutil>=5.9.0         # System monitoring

# Testing and Development
pytest>=7.4.0         # Testing framework
pytest-asyncio>=0.21.0  # Async testing
pytest-cov>=4.1.0     # Coverage testing
pytest-benchmark>=4.0.0  # Performance benchmarks
hypothesis>=6.79.0     # Property-based testing

# Performance and Timing
cProfile>=3.11.0      # Performance profiling
line_profiler>=4.0.0  # Line-by-line profiling
memory_profiler>=0.60.0  # Memory profiling
py-spy>=0.3.14        # Sampling profiler

# Mathematical and Scientific Computing
numpy>=1.24.0         # Numerical computing
scipy>=1.10.0         # Scientific computing
matplotlib>=3.7.0     # Plotting and visualization
seaborn>=0.12.0       # Statistical visualization

# Configuration and Environment
pyyaml>=6.0           # YAML configuration
python-dotenv>=1.0.0  # Environment variables
configparser>=5.3.0   # Configuration parsing
toml>=0.10.2          # TOML configuration

# Utilities
click>=8.1.0          # Command-line interface
rich>=13.4.0          # Rich text and beautiful formatting
tqdm>=4.65.0          # Progress bars
colorlog>=6.7.0       # Colored logging

# Docker and Container Support
docker>=6.1.0         # Docker Python API
docker-compose>=1.29.0  # Docker Compose support

# Development Tools
black>=23.3.0         # Code formatting
flake8>=6.0.0         # Linting
mypy>=1.3.0           # Static type checking
pre-commit>=3.3.0     # Git hooks

# Documentation
sphinx>=7.0.0         # Documentation generation
sphinx-rtd-theme>=1.2.0  # Read the Docs theme
myst-parser>=2.0.0    # Markdown parsing for Sphinx

# Security and Hardening
bandit>=1.7.5         # Security linting
safety>=2.3.0         # Dependency vulnerability checking

# Platform-specific requirements
# macOS specific (for the development environment)
pyobjc-core>=9.0.0; sys_platform == "darwin"
pyobjc-framework-CoreFoundation>=9.0.0; sys_platform == "darwin"

# Optional: Hardware acceleration support
# These may require specific hardware/drivers
pycuda>=2022.2.0; extra == "gpu"
opencl>=2022.2.0; extra == "opencl"

# Development dependencies (install with: pip install -e ".[dev]")
[dev]
jupyter>=1.0.0
ipython>=8.14.0
notebook>=6.5.4
jupyterlab>=4.0.0

# Testing dependencies (install with: pip install -e ".[test]")
[test]
pytest-xdist>=3.3.0      # Parallel testing
pytest-mock>=3.11.0      # Mocking
pytest-timeout>=2.1.0    # Test timeouts
coverage>=7.2.0          # Coverage analysis

# Production dependencies (install with: pip install -e ".[prod]")
[prod]
gunicorn>=21.0.0         # WSGI HTTP Server
uvicorn>=0.22.0          # ASGI server
supervisor>=4.2.0        # Process control
systemd-python>=234      # Systemd integration
